Aufraeumen
* Verlagerung von Würfel und Chat Ausgabefunktionen nach roll-helper.mjs, um den Zugriff von verschiedenen Stellen aus zu ermöglichen
* Vorbereitung um in Module wie Token Action HUD eingebunden werden zu können
